made DefaultMrrtReportTemplateParser package private
