--- api/src/main/java/org/openmrs/module/radiology/report/template/Parser.java
@@ -1,21 +1,0 @@
-/**
- * This Source Code Form is subject to the terms of the Mozilla Public License,
- * v. 2.0. If a copy of the MPL was not distributed with this file, You can
- * obtain one at http://mozilla.org/MPL/2.0/. OpenMRS is also distributed under
- * the terms of the Healthcare Disclaimer located at http://openmrs.org/license.
- * Copyright (C) OpenMRS Inc. OpenMRS is a registered trademark and the OpenMRS
- * graphic logo is a trademark of OpenMRS Inc.
- */
-package org.openmrs.module.radiology.report.template;
-
-import java.io.File;
-import java.io.IOException;
-import java.io.InputStream;
-
-public interface Parser {
-    
-    
-    public MrrtReportTemplate parse(File file) throws IOException;
-    
-    public MrrtReportTemplate parse(InputStream in) throws IOException;
-}
